refactor(gsd): narrow auto-worktree utility test seams
