+++ src/resources/extensions/gsd/auto-worktree-branch-lifecycle.ts
@@ -1,0 +1,6 @@
+// gsd-pi — Narrow auto-worktree branch lifecycle seam.
+//
+// Keeps branch start-point callers off the legacy auto-worktree compatibility
+// barrel while extraction continues.
+
+export { _resolveAutoWorktreeStartPoint } from "./auto-worktree.js";
