+++ src/resources/extensions/gsd/auto-worktree-path-utils.ts
@@ -1,0 +1,10 @@
+// gsd-pi — Narrow auto-worktree path utility seam.
+//
+// Keeps path comparison, pathspec, and expected-warning callers off the legacy
+// auto-worktree compatibility barrel while extraction continues.
+
+export {
+  _gitPathspecForWorktreePath,
+  _isExpectedWorktreeUnlinkError,
+  _isSamePath,
+} from "./auto-worktree.js";
