+++ src/resources/extensions/gsd/auto-worktree-runtime-cleanup.ts
@@ -1,0 +1,6 @@
+// gsd-pi — Narrow auto-worktree runtime cleanup seam.
+//
+// Keeps runtime cleanup callers off the legacy auto-worktree compatibility
+// barrel while extraction continues.
+
+export { cleanStaleRuntimeUnits } from "./auto-worktree.js";
